Arreglo error de tipo para 'cantidad_opciones'

+++ TP2/src/menu.c
@@ -9,7 +9,7 @@ typedef struct opcion {
 struct menu {
 	menu_t *menu_padre;
 	opcion_t *primer_opcion;
-	int cantidad_opciones;
+	size_t cantidad_opciones;
 };
 
 struct iterador_menu {
